+++ lib/gooddata/models/membership.rb
@@ -63,8 +63,8 @@ module GoodData
         c.create(self, json)
       end
 
-      def diff_list(list_1, list_2)
-        GoodData::Helpers.diff(list_1, list_2, key: :login)
+      def diff_list(list1, list2)
+        GoodData::Helpers.diff(list1, list2, key: :login)
       end
     end