+++ lib/gooddata/rest/client.rb
@@ -367,7 +367,7 @@ module GoodData
         @connection.download source_relative_path, target_file_path, options
       end
 
-      def download_from_user_webdav(source_relative_path, target_file_path, options = { client: GoodData.client, project: project })
+      def download_from_user_webdav(source_relative_path, target_file_path, options = { client: GoodData.client })
         download(source_relative_path, target_file_path, options.merge(:directory => options[:directory],
                                                                        :staging_url => user_webdav_path))
       end