+++ spec/integration/blueprint_updates_spec.rb
@@ -49,7 +49,7 @@ describe 'Create project using GoodData client', :constraint => 'slow' do
         reference: 'attr.commits.factsof',
         name: 'anchor_label')
     end
-    @project.update_from_blueprint(bp, preference: { cascade_drops: false,  preserve_data: false})
+    @project.update_from_blueprint(bp, maql_replacements: { "PRESERVE DATA" => ""})
     data = [
       ['label.commits.factsof.id', 'fact.commits.lines_changed', 'committed_on', 'dataset.devs', 'dataset.repos'],
       ['111', 1, '01/01/2011', '1', '1'],